requirements.txt
Pillow==10.0.1
pytest==7.4.2
pytest-cov==4.0.0
pytest-html==2.0.1